requirements.txt
google-generativeai
AppOpener
pyttsx3
SpeechRecognition
wikipedia
pywhatkit
pygetwindow
requests
pyaudio
setuptools
pycaw
firebase-admin
pyrebase4
PyQt5
